+++ src/util/csvDocument.js
@@ -4,7 +4,7 @@ import {extractFileName} from "./util";
 import {plotLoading, plotRadar} from "./factory";
 import InputSanitizer from "./inputSanitizer";
 
-export default class CSVDocument {
+export default class CsvDocument {
     constructor(url) {
         this._url = url;
         plotLoading();
